+++ releng/release-notes-next/sbom-generator.feature
@@ -5,4 +5,5 @@ full audit traceability by linking built RPMs with their original source
 tarballs and patches, including SHA-256 hashes. Supporting both CycloneDX 1.5 
 and SPDX 2.3 formats, the plugin leverages a chroot-native analysis model to 
 ensure high metadata accuracy for cross-distribution builds and compatibility 
-with modern security scanners.
+with modern security scanners, File Integrity Monitoring (FIM), and Supply Chain 
+forensic analysis.
